fix: 헤더 변경사항 적용
