+++ src/app/(with-public)/layout.tsx
@@ -1,5 +1,5 @@
 import DefaultLayout from "@/components/layout/DefaultLayout";
-import Header from "@/components/layout/header/Header";
+import Header from "@/components/layout/Header";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
    return (
