+++ src/components/layout/Header.tsx
@@ -1,8 +1,13 @@
 "use client";
 
+import HeaderUI from "@/components/layout/HeaderUI";
+import HeaderSideBarMenu from "@/components/layout/HeaderSideBarMenu";
+import { useState } from "react";
 import { usePathname } from "next/navigation";
 
 export default function Header({ children }: { children?: React.ReactNode }) {
+   const [isMenuOpen, setIsMenuOpen] = useState(false); // 헤더 모바일 메뉴
+
    const pathname = usePathname();
    const isRequestPage =
       pathname.startsWith("/request") ||
@@ -11,29 +16,21 @@ export default function Header({ children }: { children?: React.ReactNode }) {
       pathname.startsWith("/received-requests") ||
       pathname === "/my-quotes";
 
-   const headerClass =
-      "sticky top-0 left-0 z-50 bg-white border-b border-line-100";
-
-   const commonContainerClass =
-      "flex items-center max-w-[1400px] mx-auto px-6 md:px-16 lg:px-0";
-
-   const containerClass = `
-    ${commonContainerClass}
-    ${isRequestPage ? "min-h-14 lg:min-h-22" : "h-14 lg:h-22"}
-  `;
-
    return (
-      <header className={headerClass}>
-         <div className={containerClass}>
-            {isRequestPage ? (
-               <div>
-                  <div>성경님 헤더</div>
-                  <div className="py-3 lg:py-4">{children}</div>
+      <>
+         {!isRequestPage ? (
+            <HeaderUI onToggleMenu={() => setIsMenuOpen(!isMenuOpen)} />
+         ) : (
+            <div>
+               <HeaderUI onToggleMenu={() => setIsMenuOpen(!isMenuOpen)} />
+               <div className="mx-auto flex min-h-14 max-w-350 items-center bg-white px-6 py-3 md:px-16 lg:min-h-22 lg:px-0 lg:py-4">
+                  {children}
                </div>
-            ) : (
-               <div>성경님 헤더</div>
-            )}
-         </div>
-      </header>
+            </div>
+         )}
+         {isMenuOpen && (
+            <HeaderSideBarMenu onClick={() => setIsMenuOpen(false)} />
+         )}
+      </>
    );
 }
